+++ src/api/v2/ontology/ontologies-endpoint.spec.ts
@@ -4,8 +4,8 @@ import {KnoraApiConnection} from '../../../knora-api-connection';
 import {MockAjaxCall} from '../../../../test/mockajaxcall';
 import {UsersResponse} from "../../../models/admin/users-response";
 import {ReadOntology} from "../../../models/v2/ontologies/read-ontology";
-import {ResourceClassDefinition} from '../../../models/v2/ontologies/class-definition';
 import {SystemPropertyDefinition} from '../../../models/v2/ontologies/system-property-definition';
+import {ResourceClassDefinition} from '../../../models/v2/ontologies/resource-class-definition';
 
 describe('OntologiesEndpoint', () => {
